rename: main 페이지 구조변경에 맞춰 수정 및 import 경로 수정
